fix: refactor components folder
