+++ src/components/Home/ChoiceSlide/ChoiceSlide.styles.tsx
