+++ src/components/Home/Timer/Timer.tsx
